requirements.txt
groq==0.4.2
tavily-python==0.3.3
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
